+++ hata/ext/slash/form_submit_command.py
@@ -242,7 +242,7 @@ class FormSubmitCommand(CustomIdBasedCommand):
         command_coroutine = self._command_function(*positional_parameters, **keyword_parameters)
         
         try:
-            await process_command_coroutine(client, interaction_event, False, command_coroutine)
+            await process_command_coroutine(client, interaction_event, None, command_coroutine)
         except BaseException as err:
             exception = err
         else: